Improve various API weirdness (breaking change!)
- Update oxen-encoding to 1.4.0.
- Update to no longer use some of the span weird bits dropped from
  oxen-encoding in 1.4.0.
- Drop bspan, uspan, cspan typedefs.  They are a bit misleading because
  they hide an implicit const, and also hide that they are actually a
  std::span type.
- Drop connection_interface and just use Connection directly instead.
  connection_interface is now just a deprecated typedef for Connection,
  so as to hopefully not break too much current stream use.
- Combine DatagramIO and dgram_interface into a single Datagrams class.
- Revamp the datagram interface for sending datagrams so that sending a
  datagram now requires obtaining a shared_ptr<Datagrams> from the
  connection, which then can be used (sort of) like a Stream.  As this
  commit is breaking lots of API already, this also drops the old
  `send_datagram()` method of Connection, and removes
  dgram_interface::reply which didn't make any sense as a name anyway.
- Replaced the interface for receiving datagrams: instead of a callback
  getting a dgram_interface and a std::vector, you now get a `datagrams`
  struct which contains Connection/Datagrams references, a span of the
  data, and *possibly* a vector owning the data.  Previously in order to
  pass the vector we would have to copy it (except in the case of large,
  split datagrams where we had a vector), which is wasteful if the
  callback doesn't intend to preserve it.

  This most definitely breaks backwards compat, but currently that only
  affects dev Lokinet so I'm okay with it.
- Merge messages.hpp into datagrams.hpp.  The old name made no sense (it
  is basically just part of the datagram handling implementation)
- Merge src/gnutls_creds.cpp, src/gnutls_crypto.cpp, and
  src/gnutls_session.cpp and just src/gnutls_crypto.cpp to match the
  header name that defines everything in these three source files.  If
  we want to split them up again in the future, fine, but any such split
  needs to start in the header.
- Drop fully templated methods buffer-taking methods (send, respond)
  with untemplated and specific specializations to make them more useful
  w.r.t. implicit conversions from span-compatible values.
- Drop non-char basic_string and basic_string_view suport.
- combine bt req ".span" and ".body()" methods into one single "body()"
  that returns a string_view or span depending on what char type you ask
  for.
- tests: replace unnecessary bspan usage with string_views, and replace
  all REQUIRES_THAT(..., EqualsSpan) with a simpler (and nicer, if it
  fails) `REQUIRES(view(...) == )`

(This leaves the version un-bumped because it looks like dev already has
a not-yet released version bump to 1.3.0)
