+++ include/oxen/quic.hpp
@@ -12,7 +12,6 @@
 #include "quic/iochannel.hpp"
 #include "quic/ip.hpp"
 #include "quic/loop.hpp"
-#include "quic/messages.hpp"
 #include "quic/network.hpp"
 #include "quic/opt.hpp"
 #include "quic/result.hpp"
