+++ include/oxen/quic/connection.hpp
@@ -33,28 +33,82 @@
 
 namespace oxen::quic
 {
-    struct dgram_interface;
     struct IOContext;
-    class DatagramIO;
+    class Datagrams;
     struct Packet;
     class Endpoint;
     class Network;
     class Loop;
+    namespace dgram
+    {
+        struct rotating_buffer;
+    }
 
     inline constexpr uint64_t MAX_ACTIVE_CIDS{4};
     inline constexpr size_t NGTCP2_RETRY_SCIDLEN{18};
 
-    class connection_interface : public std::enable_shared_from_this<connection_interface>
+    // called when a connection's handshake completes
+    // the server will call this when it sends the final handshake packet
+    // the client will call this when it receives that final handshake packet
+    using connection_established_callback = std::function<void(Connection& conn)>;
+
+    // called when a connection closes or times out before the handshake completes
+    using connection_closed_callback = std::function<void(Connection& conn, uint64_t ec)>;
+
+    class Connection : public std::enable_shared_from_this<Connection>
     {
-      protected:
-        virtual std::shared_ptr<Stream> queue_incoming_stream_impl(
-                std::function<std::shared_ptr<Stream>(Connection& c, Endpoint& e)> make_stream) = 0;
-        virtual std::shared_ptr<Stream> open_stream_impl(
-                std::function<std::shared_ptr<Stream>(Connection& c, Endpoint& e)> make_stream) = 0;
-        virtual std::shared_ptr<Stream> get_stream_impl(int64_t id) = 0;
+        friend class Endpoint;
+        friend class Stream;
+        friend class Datagrams;
+        friend class TestHelper;
+        friend struct dgram::rotating_buffer;
+        friend struct connection_callbacks;
+        friend void conn_set_validated(Connection&);
 
       public:
-        virtual std::string_view selected_alpn() const = 0;
+        // Non-movable/non-copyable; you must always hold a Connection in a shared_ptr
+        Connection(const Connection&) = delete;
+        Connection& operator=(const Connection&) = delete;
+        Connection(Connection&&) = delete;
+        Connection& operator=(Connection&&) = delete;
+
+        // Construct and initialize a new inbound/outbound connection to/from a remote
+        //      ep: owning endpoints
+        //      scid: local ("primary") CID used for this connection (random for outgoing)
+        //		dcid: remote CID used for this connection
+        //      path: network path used to reach remote client
+        //      ctx: IO session dedicated for this connection context
+        //      alpns: passed directly to TLS session for handshake negotiation. The server
+        //          will select the first in the client's list it also supports, so the user
+        //          should list them in decreasing priority. If the user does not specify alpns,
+        //          the default will be set
+        //      default_handshake_timeout: the default timeout for handshaking for the endpoint
+        //          (individual connections might have this overridden via connect option).
+        //      remote_pk: optional parameter used by clients to verify the pubkey of the remote
+        //          endpoint during handshake negotiation. For servers, omit this parameter or
+        //          pass std::nullopt
+        //		hdr: optional parameter to pass to ngtcp2 for server specific details
+        static std::shared_ptr<Connection> make_conn(
+                Endpoint& ep,
+                ConnectionID rid,
+                const quic_cid& scid,
+                const quic_cid& dcid,
+                const Path& path,
+                std::shared_ptr<IOContext> ctx,
+                std::span<const std::string> alpns,
+                std::chrono::nanoseconds default_handshake_timeout,
+                std::optional<std::vector<unsigned char>> remote_pk = std::nullopt,
+                ngtcp2_pkt_hd* hdr = nullptr,
+                std::optional<ngtcp2_token_type> token_type = std::nullopt,
+                ngtcp2_cid* ocid = nullptr,
+                bool disable_mtu_discovery = false);
+
+        TLSSession* get_session() const { return tls_session.get(); }
+        TLSCreds* get_creds() const { return tls_creds.get(); }
+
+        std::span<const unsigned char> remote_key() const;
+
+        Direction direction() const { return dir; }
 
         /// Queues an incoming stream of the given StreamT type, forwarding the given arguments to
         /// the StreamT constructor.  The stream will be given the next unseen incoming connection
@@ -135,90 +189,39 @@ namespace oxen::quic
             throw std::out_of_range{"Could not find a stream with ID " + std::to_string(id)};
         }
 
-        template <oxenc::basic_char CharType>
-        void send_datagram(std::basic_string_view<CharType> data, std::shared_ptr<void> keep_alive)
-        {
-            send_datagram(str_to_bspan(data), std::move(keep_alive));
-        }
-
-        template <oxenc::basic_char Char>
-        void send_datagram(std::vector<Char>&& buf)
-        {
-            auto keep_alive = std::make_shared<std::vector<Char>>(std::move(buf));
-            auto bsp = vec_to_span<std::byte>(*keep_alive);
-            send_datagram(bsp, std::move(keep_alive));
-        }
-
-        template <oxenc::basic_char CharType>
-        void send_datagram(std::basic_string<CharType>&& data)
-        {
-            auto keep_alive = std::make_shared<std::basic_string<CharType>>(std::move(data));
-            std::basic_string_view<CharType> view{*keep_alive};
-            send_datagram(str_to_bspan(view), std::move(keep_alive));
-        }
-
-        template <oxenc::basic_char CharType>
-        void send_datagram(std::span<const CharType> data, std::shared_ptr<void> keep_alive)
-        {
-            send_datagram(span_to_span<std::byte>(data), std::move(keep_alive));
-        }
-
-        virtual void send_datagram(bspan data, std::shared_ptr<void> keep_alive) = 0;
-
-        virtual Endpoint& endpoint() = 0;
-        virtual const Endpoint& endpoint() const = 0;
-
-        virtual void set_close_quietly() = 0;
-
-        virtual bool datagrams_enabled() const = 0;
-        virtual bool packet_splitting_enabled() const = 0;
-        virtual const ConnectionID& reference_id() const = 0;
-        virtual bool is_validated() const = 0;
-        virtual Direction direction() const = 0;
-        virtual uspan remote_key() const = 0;
-        virtual bool is_inbound() const = 0;
-        virtual bool is_outbound() const = 0;
-        virtual std::string direction_str() = 0;
-
-        // Non-virtual base class wrappers for the virtual methods of the same name with _impl
-        // appended (e.g.  path_impl); these versions in the base class wrap the _impl call in a
-        // call_get to return the value synchronously.  Generally external application code should
-        // use this, and internal quic code (already in the event loop thread) should use the _impl
-        // ones directly.
-
         /// Returns the number of streams that are currently open
-        size_t num_streams_active();
+        size_t num_streams_active() const;
 
         /// Returns the number of streams that have been created but are not yet active on the
         /// connection; they will become active once the connection negotiates an increase to the
         /// maximum number of streams, *or* when an existing stream closes, opening a stream slot on
         /// the connection.
-        size_t num_streams_pending();
+        size_t num_streams_pending() const;
 
         /// Returns the maximum number of active streams that the connection currently allows.
-        uint64_t get_max_streams();
+        uint64_t get_max_streams() const;
 
         /// Returns the number of new streams that may still be activated on this connection.
-        uint64_t get_streams_available();
+        uint64_t get_streams_available() const;
 
         /// Returns a copy of the current Path in use by this connection.
-        Path path();
+        Path path() const;
 
         /// Returns a copy of the local address of the path in use by this connection.  (If you want
         /// both local and remote then prefer to call `path()` once instead of local() and remote()
         /// separately).
-        Address local();
+        Address local() const;
 
         /// Returns a copy of the remote address of the path in use by this connection.  (If you
         /// want both local and remote then prefer to call `path()` once instead of local() and
         /// remote() separately).
-        Address remote();
+        Address remote() const;
 
         /// Returns the maximum datagram size accepted by this connection.  This depends on the
         /// negotiated QUIC connection and can change over time, but will generally be somewhere in
         /// the 1150-1450 range when not using datagram splitting on the connection, or double that
         /// with datagram splitting enabled.
-        size_t get_max_datagram_size();
+        size_t get_max_datagram_size() const;
 
         /// Sets the maximum split datagram queue lookahead for coalescing split datagrams.  This is
         /// how far we look ahead in the datagram queue to find "small" pieces of split datagrams
@@ -256,8 +259,9 @@ namespace oxen::quic
         /// out-of-order partial packet delivery.  (The default split receive buffer is more than
         /// sufficient for any feasible lookahead value, but very small receive buffers with very
         /// large lookaheads might not be).
-        virtual void set_split_datagram_lookahead(int n) = 0;
-        virtual int get_split_datagram_lookahead() const = 0;
+        void set_split_datagram_lookahead(int n);
+        int get_split_datagram_lookahead() const;
+        bool packet_splitting_enabled() const { return _packet_splitting; }
 
         /// Obtains the current max datagram size *if* it has changed since the last time this
         /// method was called (or if this method has never been called), otherwise returns nullopt.
@@ -270,135 +274,58 @@ namespace oxen::quic
         /// This feature is for non-standard/exotic uses: if you don't care about changes to the
         /// size (for example, if you use splitting and know the size will always be sufficient)
         /// then you can safely never worry about this function.
-        virtual std::optional<size_t> max_datagram_size_changed() = 0;
+        std::optional<size_t> max_datagram_size_changed() const;
 
-        // WIP functions: these are meant to expose specific aspects of the internal state of connection
-        // and the datagram IO object for debugging and application (user) utilization.
-        //
-        //  last_cleared: returns the index of the last cleared bucket in the recv_buffer
-        virtual int last_cleared() const = 0;
+        /// Initiates connection closing, transmitting the given error code to the other size as the
+        /// reason for the connection closing.
+        void close_connection(uint64_t error_code = 0);
 
-        virtual void close_connection(uint64_t error_code = 0) = 0;
+        /// Returns a pointer to the Datagrams object for the connection, if this connection has
+        /// datagrams enabled, through which datagrams can be sent via the send() method.  Returns
+        /// nullptr if this connection does not support datagrams.
+        std::shared_ptr<Datagrams> datagrams();
 
-        virtual ~connection_interface();
+        /// Returns true if datagrams are enabled on this connection.
+        bool datagrams_enabled() const { return static_cast<bool>(dgrams); }
 
-      protected:
-        // Unsynchronized access methods suitable only for internal use in contexts where we know we
-        // are already inside the event loop thread.  The public APIs for these (without _impl)
-        // wraps these in a call_get to make them thread-safe.
-        virtual size_t num_streams_active_impl() const = 0;
-        virtual size_t num_streams_pending_impl() const = 0;
-        virtual uint64_t get_max_streams_impl() const = 0;
-        virtual uint64_t get_streams_available_impl() const = 0;
-        virtual const Path& path_impl() const = 0;
-        virtual const Address& local_impl() const { return path_impl().local; }
-        virtual const Address& remote_impl() const { return path_impl().remote; }
-        // Returns 0 if datagrams are not available
-        virtual size_t get_max_datagram_size_impl() = 0;
-    };
+        size_t num_streams_active_impl() const { return _streams.size(); }
+        size_t num_streams_pending_impl() const { return pending_streams.size(); }
 
-    class Connection : public connection_interface
-    {
-        friend class TestHelper;
-        friend struct rotating_buffer;
-        friend struct connection_callbacks;
-
-      public:
-        // Non-movable/non-copyable; you must always hold a Connection in a shared_ptr
-        Connection(const Connection&) = delete;
-        Connection& operator=(const Connection&) = delete;
-        Connection(Connection&&) = delete;
-        Connection& operator=(Connection&&) = delete;
-
-        // Construct and initialize a new inbound/outbound connection to/from a remote
-        //      ep: owning endpoints
-        //      scid: local ("primary") CID used for this connection (random for outgoing)
-        //		dcid: remote CID used for this connection
-        //      path: network path used to reach remote client
-        //      ctx: IO session dedicated for this connection context
-        //      alpns: passed directly to TLS session for handshake negotiation. The server
-        //          will select the first in the client's list it also supports, so the user
-        //          should list them in decreasing priority. If the user does not specify alpns,
-        //          the default will be set
-        //      default_handshake_timeout: the default timeout for handshaking for the endpoint
-        //          (individual connections might have this overridden via connect option).
-        //      remote_pk: optional parameter used by clients to verify the pubkey of the remote
-        //          endpoint during handshake negotiation. For servers, omit this parameter or
-        //          pass std::nullopt
-        //		hdr: optional parameter to pass to ngtcp2 for server specific details
-        static std::shared_ptr<Connection> make_conn(
-                Endpoint& ep,
-                ConnectionID rid,
-                const quic_cid& scid,
-                const quic_cid& dcid,
-                const Path& path,
-                std::shared_ptr<IOContext> ctx,
-                std::span<const std::string> alpns,
-                std::chrono::nanoseconds default_handshake_timeout,
-                std::optional<std::vector<unsigned char>> remote_pk = std::nullopt,
-                ngtcp2_pkt_hd* hdr = nullptr,
-                std::optional<ngtcp2_token_type> token_type = std::nullopt,
-                ngtcp2_cid* ocid = nullptr,
-                bool disable_mtu_discovery = false);
-
-        void packet_io_ready();
-
-        TLSSession* get_session() const { return tls_session.get(); }
-        TLSCreds* get_creds() const { return tls_creds.get(); }
-
-        uspan remote_key() const override;
-
-        Direction direction() const override { return dir; }
-
-        size_t num_streams_active_impl() const override { return _streams.size(); }
-        size_t num_streams_pending_impl() const override { return pending_streams.size(); }
-
-        void halt_events();
         bool is_closing() const { return closing; }
-        void set_closing() { closing = true; }
         bool is_draining() const { return draining; }
+
+        bool is_outbound() const { return _is_outbound; }
+        bool is_inbound() const { return not is_outbound(); }
+        std::string direction_str() { return is_inbound() ? "SERVER"s : "CLIENT"s; }
+
+        Endpoint& endpoint() { return _endpoint; }
+        const Endpoint& endpoint() const { return _endpoint; }
+
+        std::string_view selected_alpn() const;
+
+        size_t get_max_datagram_piece() const;
+
+        std::optional<size_t> max_datagram_size_changed();
+
+        // True if the connection has been accepted and the remote pubkey successfully verified.
+        bool is_validated() const { return _is_validated; }
+
+        const ConnectionID& reference_id() const { return _ref_id; }
+
+        void set_close_quietly();
+
+        bool closing_quietly() const { return _close_quietly; }
+
+      private:
+        void packet_io_ready();
+        void halt_events();
+        void set_closing() { closing = true; }
         void set_draining() { draining = true; }
         stream_data_callback get_default_data_callback() const;
-
-        bool is_outbound() const override { return _is_outbound; }
-        bool is_inbound() const override { return not is_outbound(); }
-        std::string direction_str() override { return is_inbound() ? "SERVER"s : "CLIENT"s; }
-
-        const Path& path_impl() const override { return _path; }
-        const Address& local_impl() const override { return _path.local; }
-        const Address& remote_impl() const override { return _path.remote; }
-
-        Endpoint& endpoint() override { return _endpoint; }
-        const Endpoint& endpoint() const override { return _endpoint; }
-
-        std::string_view selected_alpn() const override;
-
-        uint64_t get_streams_available_impl() const override;
-        size_t get_max_datagram_piece();
-        size_t get_max_datagram_size_impl() override;
-        uint64_t get_max_streams_impl() const override { return _max_streams; }
-
-        bool datagrams_enabled() const override;
-        bool packet_splitting_enabled() const override { return _packet_splitting; }
-        void set_split_datagram_lookahead(int n) override;
-        int get_split_datagram_lookahead() const override;
-
-        std::optional<size_t> max_datagram_size_changed() override;
-
-        // public debug functions; to be removed with friend test fixture class
-        int last_cleared() const override;
-
-        void send_datagram(bspan data, std::shared_ptr<void> keep_alive) override;
-
-        void close_connection(uint64_t error_code = 0) override;
 
         // This mutator is called from the gnutls code after cert verification (if it is successful)
         void set_validated();
 
-        bool is_validated() const override { return _is_validated; }
-
-        // These are public so we can access them from the ngtcp free floating functions
-        // (on_handshake_completed and on_handshake_confirmed) and when the connection is closed
         connection_established_callback conn_established_cb;
         connection_closed_callback conn_closed_cb;
 
@@ -424,11 +351,7 @@ namespace oxen::quic
 
         void set_new_path(Path new_path);
 
-        const ConnectionID& reference_id() const override { return _ref_id; }
-
-        void set_close_quietly() override;
-
-        bool closing_quietly() const { return _close_quietly; }
+        uint64_t get_streams_available_impl() const;
 
         // Called when the endpoint drops its shared pointer to this Connection, to have this
         // Connection clear itself from all of its streams and then drop the streams.  (This is a
@@ -437,7 +360,6 @@ namespace oxen::quic
         // alive after it gets dropped from libquic internal structures).
         void drop_streams();
 
-      private:
         // private Constructor (publicly construct via `make_conn` instead, so that we can properly
         // set up the shared_from_this shenanigans).
         Connection(
@@ -477,15 +399,16 @@ namespace oxen::quic
 
         const uint64_t _max_streams{DEFAULT_MAX_BIDI_STREAMS};
         const bool _packet_splitting{false};
-        size_t _last_max_dgram_piece{0};
-        std::atomic<bool> _max_dgram_size_changed{true};
+        mutable size_t _last_max_dgram_piece{0};
+        mutable std::atomic<bool> _max_dgram_size_changed{true};
 
         std::atomic<bool> _close_quietly{false};
         std::atomic<bool> _is_validated{false};
 
         std::vector<unsigned char> remote_pubkey{};
 
-        void revert_early_streams();
+        void revert_early_channels();
+        void reset_early_datagrams();
 
         struct connection_deleter
         {
@@ -525,12 +448,12 @@ namespace oxen::quic
                 std::optional<int64_t> stream_id = std::nullopt);
 
         std::shared_ptr<Stream> queue_incoming_stream_impl(
-                std::function<std::shared_ptr<Stream>(Connection& c, Endpoint& e)> make_stream) override;
+                std::function<std::shared_ptr<Stream>(Connection& c, Endpoint& e)> make_stream);
 
         std::shared_ptr<Stream> open_stream_impl(
-                std::function<std::shared_ptr<Stream>(Connection& c, Endpoint& e)> make_stream) override;
+                std::function<std::shared_ptr<Stream>(Connection& c, Endpoint& e)> make_stream);
 
-        std::shared_ptr<Stream> get_stream_impl(int64_t id) override;
+        std::shared_ptr<Stream> get_stream_impl(int64_t id);
 
         // holds a mapping of active streams
         std::map<int64_t, std::shared_ptr<Stream>> _streams;
@@ -539,7 +462,7 @@ namespace oxen::quic
         int64_t next_incoming_stream_id = is_outbound() ? 1 : 0;
 
         // datagram "pseudo-stream"
-        std::shared_ptr<DatagramIO> datagrams;
+        std::shared_ptr<Datagrams> dgrams;
         // "pseudo-stream" to represent ngtcp2 stream ID -1
         std::shared_ptr<Stream> pseudo_stream;
         // holds queue of pending streams not yet ready to broadcast
@@ -555,8 +478,6 @@ namespace oxen::quic
 
         io_result read_packet(const Packet& pkt);
 
-        std::shared_ptr<dgram_interface> di;
-
         /********* TEST SUITE FUNCTIONALITY *********/
         void set_local_addr(Address new_local);
         bool debug_datagram_drop_enabled{false};
@@ -569,12 +490,12 @@ namespace oxen::quic
         // these are public so ngtcp2 can access them from callbacks
         int stream_opened(int64_t id);
         int stream_ack(int64_t id, size_t size);
-        int stream_receive(int64_t id, bspan data, bool fin);
+        int stream_receive(int64_t id, std::span<const std::byte> data, bool fin);
         void stream_execute_close(Stream& s, uint64_t app_code);
         void stream_closed(int64_t id, uint64_t app_code);
         void close_all_streams();
         void check_pending_streams(uint64_t available);
-        int recv_datagram(bspan data, bool fin);
+        int recv_datagram(std::span<const std::byte> data, bool fin);
         int ack_datagram(uint64_t dgram_id);
         int recv_token(const uint8_t* token, size_t tokenlen);
 
@@ -594,8 +515,10 @@ namespace oxen::quic
         // externally.
         void check_stream_timeouts();
 
-        ~Connection() override;
+        ~Connection();
     };
+
+    using connection_interface [[deprecated("use Connection directly instead")]] = Connection;
 
     extern "C"
     {
