+++ include/oxen/quic/iochannel.hpp
@@ -2,7 +2,6 @@
 
 #include "address.hpp"
 #include "connection_ids.hpp"
-#include "messages.hpp"
 #include "utils.hpp"
 
 #include <oxenc/common.h>
@@ -15,7 +14,6 @@
 #include <memory>
 #include <optional>
 #include <span>
-#include <stdexcept>
 #include <string>
 #include <string_view>
 #include <type_traits>
@@ -29,7 +27,12 @@ namespace oxen::quic
     class Endpoint;
     class Loop;
     class Stream;
+    namespace dgram
+    {
+        struct prepared;
+    }
 
+    /// Base class shared by streams and datagrams.
     class IOChannel
     {
       protected:
@@ -40,6 +43,9 @@ namespace oxen::quic
 
         Endpoint& endpoint;
         Loop& loop;
+
+        // The fixed Connection reference_id.  This will be the same as `get_conn()->reference_id`
+        // while the connection exists, but persists even if the connection object gets destroyed.
         const ConnectionID reference_id;
 
         // no copy, no move. always hold in a shared pointer
@@ -52,6 +58,10 @@ namespace oxen::quic
         virtual std::shared_ptr<Stream> get_stream() = 0;
         virtual int64_t stream_id() const = 0;
 
+        // Returns this channel's connection object.  Will return a nullptr if the Connection no
+        // longer exists.
+        std::shared_ptr<Connection> get_conn();
+
         // These public methods are intended for access from anywhere and invoke a call_get to
         // synchronously get their value.  (Subclasses provide implementations by overriding the
         // protected _impl versions of these methods).
@@ -60,39 +70,28 @@ namespace oxen::quic
         bool has_unsent() const;
         bool is_closing() const;
 
-        // These are call_get-proxied to return the value from the Connection object.  They throw if
-        // the Connection object no longer exists.
-        Path path() const;
-        Address local() const;
-        Address remote() const;
-
-        void send(bspan data, std::shared_ptr<void> keep_alive) { send_impl(data, std::move(keep_alive)); }
-
-        template <oxenc::basic_char CharType>
-        void send(std::span<const CharType> data, std::shared_ptr<void> keep_alive)
+        void send(std::span<const std::byte> data, std::shared_ptr<void> keep_alive)
         {
-            send_impl(span_to_span<std::byte>(data), std::move(keep_alive));
+            send_impl(data, std::move(keep_alive));
         }
 
-        template <oxenc::basic_char CharType>
-        void send(std::basic_string_view<CharType> data, std::shared_ptr<void> keep_alive)
+        void send(std::span<const unsigned char> data, std::shared_ptr<void> keep_alive)
         {
-            send_impl(str_to_bspan(data), std::move(keep_alive));
+            send_impl(reinterpret_span<const std::byte>(data), std::move(keep_alive));
         }
 
-        template <oxenc::basic_char CharType>
-        void send(std::basic_string<CharType>&& data)
+        void send(std::string_view data, std::shared_ptr<void> keep_alive)
         {
-            auto keep_alive = std::make_shared<std::basic_string<CharType>>(std::move(data));
-            std::basic_string_view<CharType> view{*keep_alive};
-            send_impl(str_to_bspan(view), std::move(keep_alive));
+            send_impl(reinterpret_span<const std::byte>(data), std::move(keep_alive));
         }
+
+        void send(std::string&& data);
 
         template <oxenc::basic_char Char>
         void send(std::vector<Char>&& buf)
         {
             auto keep_alive = std::make_shared<std::vector<Char>>(std::move(buf));
-            auto bsp = vec_to_span<std::byte>(*keep_alive);
+            auto bsp = reinterpret_span<const std::byte>(std::span{*keep_alive});
             send_impl(bsp, std::move(keep_alive));
         }
 
@@ -104,10 +103,10 @@ namespace oxen::quic
 
         // This is the (single) send implementation that implementing classes must provide; other
         // calls to send are converted into calls to this.
-        virtual void send_impl(bspan, std::shared_ptr<void> keep_alive) = 0;
+        virtual void send_impl(std::span<const std::byte>, std::shared_ptr<void> keep_alive) = 0;
 
         virtual std::vector<ngtcp2_vec> pending() = 0;
-        virtual std::optional<prepared_datagram> pending_datagram(bool) = 0;
+        virtual std::optional<dgram::prepared> pending_datagram(bool) = 0;
         virtual bool sent_fin() const = 0;
         virtual void set_fin(bool) = 0;
         virtual void wrote(size_t) = 0;
@@ -123,28 +122,15 @@ namespace oxen::quic
         // Wraps an IOChannel (or derived type) accessor member function pointer in a call_get for
         // synchronous access that always returns by value (even if the member function returns by
         // reference).
-        template <typename Class, typename T, typename Ret = std::remove_cvref_t<T>, typename EP = Endpoint>
-            requires std::derived_from<Class, IOChannel>
+        template <
+                std::derived_from<IOChannel> Class,
+                typename T,
+                typename Ret = std::remove_cvref_t<T>,
+                typename EP = Endpoint>
         Ret call_get_accessor(T (Class::*getter)() const) const
         {
             return static_cast<EP&>(endpoint).loop.call_get(
                     [this, &getter]() -> Ret { return (static_cast<const Class*>(this)->*getter)(); });
-        }
-
-        // Wraps a Connection accessor member function pointer in a call_get for synchronous access
-        // that always returns by value (even if the member function returns by reference) through
-        // the IOChannel's connection pointer.  Throws if the Connection doesn't exist anymore.
-        template <typename T, typename Ret = std::remove_cvref_t<T>, typename EP = Endpoint>
-        Ret call_get_accessor(T (Connection::*getter)() const) const
-        {
-            // This do-nothing static cast to force deferred instantiation until later on (when the
-            // Endpoint class will be available).  Otherwise this won't compile because Endpoint
-            // is only forward declared here.
-            return static_cast<EP&>(endpoint).loop.call_get([this, &getter]() -> Ret {
-                if (!_conn)
-                    throw std::runtime_error{"Connection has gone away"};
-                return (_conn->*getter)();
-            });
         }
     };
