+++ include/oxen/quic/udp.hpp
@@ -36,7 +36,7 @@ namespace oxen::quic
     {
         Path path;
         ngtcp2_pkt_info pkt_info{};
-        std::variant<bspan, std::vector<std::byte>> pkt_data;
+        std::variant<std::span<const std::byte>, std::vector<std::byte>> pkt_data;
 
         size_t size() const
         {
@@ -49,7 +49,9 @@ namespace oxen::quic
         std::span<const Char> data() const
         {
             return std::visit(
-                    [](const auto& d) { return std::span<const Char>{reinterpret_cast<const Char*>(d.data()), d.size()}; },
+                    [](const auto& d) {
+                        return std::span<const Char>{reinterpret_cast<const Char*>(d.data()), d.size()};
+                    },
                     pkt_data);
         }
 
@@ -60,14 +62,14 @@ namespace oxen::quic
         void ensure_owned_data();
 
         /// Constructs a packet from a path and data view:
-        Packet(Path p, bspan d) : path{std::move(p)}, pkt_data{d} {}
+        Packet(Path p, std::span<const std::byte> d) : path{std::move(p)}, pkt_data{d} {}
 
         /// Constructs a packet from a path and transferred data:
         Packet(Path p, std::vector<std::byte>&& d) : path{std::move(p)}, pkt_data{std::move(d)} {}
 
         /// Constructs a packet from a local address, data, and the IP header; remote addr and ECN
         /// data are extracted from the header.
-        Packet(const Address& local, bspan data, msghdr& hdr);
+        Packet(const Address& local, std::span<const std::byte> data, msghdr& hdr);
     };
 
     /// RAII class wrapping a UDP socket; the socket is bound at construction and closed during
@@ -139,7 +141,7 @@ namespace oxen::quic
         ~UDPSocket();
 
       private:
-        void process_packet(bspan payload, msghdr& hdr);
+        void process_packet(std::span<const std::byte> payload, msghdr& hdr);
         io_result receive();
 
         socket_t sock_;
