+++ tests/006-server-send.cpp
@@ -7,8 +7,7 @@ namespace oxen::quic::test
     TEST_CASE("006 - Server streams: Direct creation and transmission", "[006][server][streams][send][execute]")
     {
         Network test_net{};
-        auto msg_str = "hello from the other siiiii-iiiiide"sv;
-        auto msg = to_span<std::byte>(msg_str);
+        constexpr auto msg = "hello from the other siiiii-iiiiide"sv;
 
         std::atomic<int> data_check{0};
 
@@ -25,13 +24,13 @@ namespace oxen::quic::test
             return 0;
         };
 
-        stream_data_callback server_io_data_cb = [&](IOChannel&, bspan) {
+        stream_data_callback server_io_data_cb = [&](IOChannel&, std::span<const std::byte>) {
             log::debug(test_cat, "Calling server stream data callback... data received... incrementing counter...");
             data_check += 1;
             server_promise.set_value();
         };
 
-        stream_data_callback client_io_data_cb = [&](IOChannel&, bspan) {
+        stream_data_callback client_io_data_cb = [&](IOChannel&, std::span<const std::byte>) {
             log::debug(test_cat, "Calling client stream data callback... data received... incrementing counter...");
             data_check += 1;
             client_promise.set_value();
@@ -65,16 +64,14 @@ namespace oxen::quic::test
     TEST_CASE("006 - Server streams: Remote initiation, server send", "[006][server][streams][send][execute]")
     {
         Network test_net{};
-        auto msg_str = "hello from the other siiiii-iiiiide"sv;
-        auto response_str = "okay okay i get it already"sv;
-        auto msg = to_span<std::byte>(msg_str);
-        auto response = to_span<std::byte>(response_str);
+        constexpr auto msg = "hello from the other siiiii-iiiiide"sv;
+        constexpr auto response = "okay okay i get it already"sv;
 
         std::atomic<int> ci{0}, si{0};
         std::atomic<int> data_check{0};
 
         std::shared_ptr<Stream> server_extracted_stream, client_extracted_stream;
-        std::shared_ptr<connection_interface> server_ci;
+        std::shared_ptr<Connection> server_ci;
 
         std::vector<std::promise<void>> server_promises{3}, client_promises{3};
         std::vector<std::future<void>> server_futures{3}, client_futures{3};
@@ -115,7 +112,7 @@ namespace oxen::quic::test
             return 0;
         };
 
-        stream_data_callback server_io_data_cb = [&](Stream&, bspan) {
+        stream_data_callback server_io_data_cb = [&](Stream&, std::span<const std::byte>) {
             log::debug(test_cat, "Calling server stream data callback... data received... incrementing counter...");
             data_check += 1;
             try
@@ -129,7 +126,7 @@ namespace oxen::quic::test
             }
         };
 
-        stream_data_callback client_io_data_cb = [&](Stream&, bspan) {
+        stream_data_callback client_io_data_cb = [&](Stream&, std::span<const std::byte>) {
             log::debug(test_cat, "Calling client stream data callback... data received... incrementing counter...");
             data_check += 1;
             try
