+++ tests/008-conn_hooks.cpp
@@ -9,11 +9,11 @@ namespace oxen::quic::test
         uint64_t error_code = 12345;
         uint64_t client_error{0};
         uint64_t server_error{0};
-        auto client_established = callback_waiter{[](connection_interface&) {}};
-        auto server_established = callback_waiter{[](connection_interface&) {}};
+        auto client_established = callback_waiter{[](Connection&) {}};
+        auto server_established = callback_waiter{[](Connection&) {}};
         // this needs to be destroyed *after* Network, as it may be called during ~Network
-        auto client_closed = callback_waiter{[&client_error](connection_interface&, uint64_t ec) { client_error = ec; }};
-        auto server_closed = callback_waiter{[&server_error](connection_interface&, uint64_t ec) { server_error = ec; }};
+        auto client_closed = callback_waiter{[&client_error](Connection&, uint64_t ec) { client_error = ec; }};
+        auto server_closed = callback_waiter{[&server_error](Connection&, uint64_t ec) { server_error = ec; }};
 
         Network test_net{};
